Construção dos endpoints para produtos findAll e findById e suas impl. no service

+++ devstore_backend/src/main/java/com/domain/devstore_backend/services/ProductService.java
@@ -1,6 +1,5 @@
 package com.domain.devstore_backend.services;
 
-import com.domain.devstore_backend.dto.ProductDto;
 import com.domain.devstore_backend.entities.Product;
 
 import java.util.List;
@@ -9,11 +8,11 @@ public interface ProductService {
 
     List<Product> findAll();
 
-    Product findById(Long id);
+    Product findById(Integer id);
 
     Product create(Product product);
 
-    Product update(Long id, Product product);
+    Product update(Integer id, Product product);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
